@@ -8,11 +8,11 @@
         <span class="text-gray-300 dark:text-gray-500">{{ lastRoll }}</span>
         <span>{{ currentRoll }}</span>
       </div>
-      
+
       <ButtonComponent
         type="primary"
         class="w-full my-2"
-        @click="roll"
+        @click="actionsStore.roll"
       >
         Roll
       </ButtonComponent>
@@ -24,29 +24,15 @@
   </CardComponent>
 </template>
 
-<script>
-import CardComponent from './CardComponent';
-import ButtonComponent from './ButtonComponent';
-import HeadingComponent from './HeadingComponent';
-import SaveMenuComponent from './SaveMenuComponent';
-import LoadMenuComponent from './LoadMenuComponent';
-import { mapActions, mapGetters, mapState } from 'vuex';
+<script setup>
+import CardComponent from "./CardComponent.vue";
+import ButtonComponent from "./ButtonComponent.vue";
+import HeadingComponent from "./HeadingComponent.vue";
+import SaveMenuComponent from "./SaveMenuComponent.vue";
+import LoadMenuComponent from "./LoadMenuComponent.vue";
+import { storeToRefs } from "pinia";
+import { useActionsStore } from "Stores/actions";
 
-export default {
-  name: 'ActionsPane',
-  components: {
-    ButtonComponent,
-    CardComponent,
-    HeadingComponent,
-    SaveMenuComponent,
-    LoadMenuComponent,
-  },
-  computed: {
-    ...mapGetters('actions', ['die', 'currentRoll']),
-    ...mapState('actions', ['d6', 'd10', 'lastRoll']),
-  },
-  methods: {
-    ...mapActions('actions', ['roll']),
-  }
-}
+const actionsStore = useActionsStore();
+const { lastRoll, currentRoll } = storeToRefs(actionsStore);
 </script>
