@@ -1,11 +1,14 @@
 <template>
   <CardComponent>
     <div class="flex border-b mb-2">
-      <HeadingComponent :class="{'flex-1': true, 'line-through': memory.forgotten}" level="6">
+      <HeadingComponent
+        :class="{ 'flex-1': true, 'line-through': memory.forgotten }"
+        level="6"
+      >
         {{ memory.description }}
       </HeadingComponent>
       <div class="flex-initial text-right">
-        <span 
+        <span
           class="cursor-pointer mx-2 hover:text-gray-400"
           @click="$emit('edit-memory', memory)"
         >
@@ -13,34 +16,36 @@
         </span>
       </div>
     </div>
-    
+
     <ul class="my-3">
-      <li
-          v-for="event in events(memory)"
-          :key="`event-${event.id}`"
-      >
-          <CardComponent>
-            <div class="flex">
-              <span :class="{'flex-1': true, 'line-through': memory.forgotten, 'text-gray-400': memory.forgotten}">{{event.description}}</span>
-              <div class="text-right flex-inital">
-                <RemoveCrossComponent @remove="$emit('remove-event', event)" />
-              </div>
+      <li v-for="event in events(memory)" :key="`event-${event.id}`">
+        <CardComponent>
+          <div class="flex">
+            <span
+              :class="{
+                'flex-1': true,
+                'line-through': memory.forgotten,
+                'text-gray-400': memory.forgotten,
+              }"
+              >{{ event.description }}</span
+            >
+            <div class="text-right flex-inital">
+              <RemoveCrossComponent @remove="$emit('remove-event', event)" />
             </div>
-          </CardComponent>
+          </div>
+        </CardComponent>
       </li>
     </ul>
-    
-    <FormToggleComponent 
+
+    <FormToggleComponent
       @save="add"
       @toggle="toggleControls"
       :showControls="showControls"
       v-if="canAddEvents"
     >
-      <template #button>  
-        Add an Event?
-      </template>
+      <template #button> Add an Event? </template>
       <template #form>
-        <input 
+        <input
           type="text"
           placeholder="Description"
           class="shadow appearance-none border rounded w-full py-1 px-2 m-1 text-gray-700 leading-tight focus:outline-none focus:ring-2 ring-gray-200"
@@ -49,33 +54,31 @@
         />
       </template>
     </FormToggleComponent>
-  
 
-    <div class="my-2 grid grid-rows gap-2" v-show="canToggle && ((!memory.forgotten || memory.forgotten && memory.diary !== '') || canAddMemories)">
-      <ButtonComponent
-        class="w-full"
-        @click="$emit('toggle-memory', memory)"
-      >
-        <span v-if="memory.forgotten">
-          Recover Memory
-        </span>
-        <span v-else-if="memory.diary === ''">
-          Forget Memory
-        </span>
-        <span v-else>
-          Scratch Out Memory
-        </span>
+    <div
+      class="my-2 grid grid-rows gap-2"
+      v-show="
+        canToggle &&
+        (!memory.forgotten ||
+          (memory.forgotten && memory.diary !== '') ||
+          canAddMemories)
+      "
+    >
+      <ButtonComponent class="w-full" @click="$emit('toggle-memory', memory)">
+        <span v-if="memory.forgotten"> Recover Memory </span>
+        <span v-else-if="memory.diary === ''"> Forget Memory </span>
+        <span v-else> Scratch Out Memory </span>
       </ButtonComponent>
-      
+
       <template v-if="canDiarise">
-        <ButtonComponent 
+        <ButtonComponent
           class="w-full"
           @click="$emit('diarise-memory', memory)"
           v-if="memory.diary === '' && !isDiaryFull"
         >
           Send to Diary
         </ButtonComponent>
-        <ButtonComponent 
+        <ButtonComponent
           class="w-full"
           @click="$emit('undiarise-memory', memory)"
           v-else-if="memory.diary !== '' && canAddMemories"
@@ -87,78 +90,80 @@
   </CardComponent>
 </template>
 
-<script>
-import CardComponent from 'Components/CardComponent';
-import ButtonComponent from 'Components/ButtonComponent';
-import FormToggleComponent from 'Components/FormToggleComponent';
-import HeadingComponent from 'Components/HeadingComponent';
-import RemoveCrossComponent from 'Components/RemoveCrossComponent';
-import { mapMutations, mapActions, mapGetters } from 'vuex';
-import { eventEntityFactory } from 'Libs/entities/memories';
+<script setup>
+import CardComponent from "Components/CardComponent";
+import ButtonComponent from "Components/ButtonComponent";
+import FormToggleComponent from "Components/FormToggleComponent";
+import HeadingComponent from "Components/HeadingComponent";
+import RemoveCrossComponent from "Components/RemoveCrossComponent";
+import { eventEntityFactory } from "Libs/entities/memories";
+import { useMemoriesStore } from "Stores/memories";
+import { useResourcesStore } from "Stores/resources";
+import { useNotificationsStore } from "Stores/notifications";
+import { computed, ref } from "vue";
 
-export default {
-  name: 'MemoryComponent',
-  props: {
-    memory: {
-      type: Object,
-      required: true,
-      validator: (memory) => 'string' === typeof memory.diary && 'boolean' === typeof memory.forgotten,
-    },
-    canAddMemories: {
-      type: Boolean,
-      required: true,
-    },
-    canAddEvents: {
-      type: Boolean,
-      required: true,
-    },
-    canDiarise: {
-      type: Boolean,
-      required: true,
-    },
-    canToggle: {
-      type: Boolean,
-      required: true,
-    },
-  },
-  data: function() {
-      return {
-          showControls: false,
-          newEvent:  eventEntityFactory({memory: this.memory.id}),
-      }
-  },
-  components: {
-    ButtonComponent,
-    CardComponent,
-    FormToggleComponent,
-    HeadingComponent,
-    RemoveCrossComponent,
-  },
-  computed: {
-    ...mapGetters('resources', ['hasDiary', 'isDiaryFull']),
-    ...mapGetters('memories', ['events']),
-  },
-  methods: {
-    ...mapMutations('notifications', {
-      hideNotification: 'hide'
-    }),
-    ...mapActions('notifications', ['showNotification']),
-    add(){
-      if (this.newEvent.description === '') {
-        this.showNotification({message: 'You must provide a description', type: 'warning'});
-        return;
-      }
+const emit = defineEmits([
+  "edit-memory",
+  "toggle-memory",
+  "diarise-memory",
+  "undiarise-memory",
+  "add-event",
+  "remove-event",
+]);
 
-      const event = eventEntityFactory(this.newEvent);
+const props = defineProps({
+  memory: {
+    type: Object,
+    required: true,
+    validator: (memory) =>
+      "string" === typeof memory.diary && "boolean" === typeof memory.forgotten,
+  },
+  canAddMemories: {
+    type: Boolean,
+    required: true,
+  },
+  canAddEvents: {
+    type: Boolean,
+    required: true,
+  },
+  canDiarise: {
+    type: Boolean,
+    required: true,
+  },
+  canToggle: {
+    type: Boolean,
+    required: true,
+  },
+});
 
-      this.$emit('add-event', event);
-      this.toggleControls();
-    },
-    toggleControls() {
-      this.hideNotification();
-      this.showControls = !this.showControls;
-      this.newEvent = eventEntityFactory({memory: this.memory.id});
-    },
+const memorysStore = useMemoriesStore();
+const resourcesStore = useResourcesStore();
+const notificationsStore = useNotificationsStore();
+
+const showControls = ref(false);
+const newEvent = ref(eventEntityFactory({ memory: props.memory.id }));
+
+const events = computed(() => memorysStore.sortedEvents);
+const isDiaryFull = computed(() => resourcesStore.isDiaryFull);
+
+const toggleControls = () => {
+  notificationsStore.hide();
+  showControls.value = !showControls.value;
+  newEvent.value = eventEntityFactory({ memory: props.memory.id });
+};
+
+const add = () => {
+  if (newEvent.value.description === "") {
+    notificationsStore.showNotification({
+      message: "You must provide a description",
+      type: "warning",
+    });
+    return;
   }
-}
+
+  const event = eventEntityFactory(newEvent.value);
+
+  emit("add-event", event);
+  toggleControls();
+};
 </script>
