@@ -1,11 +1,7 @@
 <template>
-  <div class="m-4 p-4 border rounded border-gray-100 shadow-md hover:shadow-xl transition-all">
+  <div
+    class="m-4 p-4 border rounded border-gray-100 shadow-md hover:shadow-xl transition-all"
+  >
     <slot />
   </div>
 </template>
-
-<script>
-export default {
-  name: 'CardComponent',
-}
-</script>
