@@ -1,14 +1,12 @@
 <template>
-    <span 
-        class="cursor-pointer select-none mx-1 hover:text-gray-400"
-        title="Remove item"
-        @click="$emit('remove')"
-        v-html="'&times;'"
-    />
+  <span
+    class="cursor-pointer select-none mx-1 hover:text-gray-400"
+    title="Remove item"
+    @click="$emit('remove')"
+    v-html="'&times;'"
+  />
 </template>
 
-<script>
-export default {
-  name: 'RemoveCrossComponent',
-}
+<script setup>
+const emit = defineEmits(["remove"]);
 </script>
