@@ -1,67 +1,53 @@
 <template>
-    <SlideDownPanelComponent v-model="saving">
-        <template #closed-heading>
-            Save
-        </template>
-        <div class="grid grid-rows gap-1 my-2">
-            <a
-                class="hidden"
-                download='save-game'
-                target="_blank"
-                ref="download"
-                href='about:blank'
-            />
-            <ButtonComponent 
-              class="w-full"
-              @click="toFile"
-            >
-              To File
-            </ButtonComponent>
-            <ButtonComponent 
-              class="w-full"
-              @click="toLocalStorage"
-              v-if="supportsLocalStorage"
-            >
-              To Local Storage
-            </ButtonComponent>
-          </div>
-    </SlideDownPanelComponent>
+  <SlideDownPanelComponent v-model="saving">
+    <template #closed-heading> Save </template>
+    <div class="grid grid-rows gap-1 my-2">
+      <a
+        class="hidden"
+        download="save-game"
+        target="_blank"
+        ref="download"
+        href="about:blank"
+      />
+      <ButtonComponent class="w-full" @click="toFile">
+        To File
+      </ButtonComponent>
+      <ButtonComponent
+        class="w-full"
+        @click="toLocalStorage"
+        v-if="doesSupportLocalStorage"
+      >
+        To Local Storage
+      </ButtonComponent>
+    </div>
+  </SlideDownPanelComponent>
 </template>
 
-<script>
-import ButtonComponent from 'Components/ButtonComponent';
-import SlideDownPanelComponent from 'Components/SlideDownPanelComponent';
-import { getStateFromStore, serialize } from 'Libs/gameState';
-import localStorage, { supportsLocalStorage } from 'Libs/localStorage';
+<script setup>
+import ButtonComponent from "Components/ButtonComponent";
+import SlideDownPanelComponent from "Components/SlideDownPanelComponent";
+import { getStateFromStore, serialize } from "Libs/gameState";
+import localStorage, { supportsLocalStorage } from "Libs/localStorage";
+import { ref, computed, useTemplateRef } from "vue";
 
-export default {
-  name: 'SaveMenuComponent',
-  data() {
-    return {
-      saving: false,
-    }
-  },
-  components: {
-    SlideDownPanelComponent,
-    ButtonComponent,
-  },
-  computed: {
-    supportsLocalStorage() {
-      return supportsLocalStorage();
-    },
-  },
-  methods: {
-    toFile() {
-      const data = serialize(getStateFromStore(this.$store));
-      this.$refs.download.href = URL.createObjectURL(new Blob([data], {type: 'text/plain'}));
-      this.$refs.download.click();
-      this.saving = false;
-    },
-    toLocalStorage() {
-      const data = serialize(getStateFromStore(this.$store));
-      localStorage.set('save-game', data);
-      this.saving = false;
-    },
-  }
-}
+const download = useTemplateRef("download");
+
+const saving = ref(false);
+
+const doesSupportLocalStorage = computed(() => supportsLocalStorage());
+
+const toFile = () => {
+  const data = serialize(getStateFromStore());
+  download.value.href = URL.createObjectURL(
+    new Blob([data], { type: "text/plain" })
+  );
+  download.value.click();
+  saving.value = false;
+};
+
+const toLocalStorage = () => {
+  const data = serialize(getStateFromStore());
+  localStorage.set("save-game", data);
+  saving.loading = false;
+};
 </script>
