@@ -1,11 +1,10 @@
-import Vue from 'vue'
-import App from './App.vue'
-import './assets/tailwind.css'
-import store from './store'
+import { createApp } from "vue";
+import { createPinia } from "pinia";
+import App from "./App.vue";
+import "./assets/tailwind.css";
 
-Vue.config.productionTip = false
+const app = createApp(App);
+const pinia = createPinia();
 
-new Vue({
-  store,
-  render: h => h(App)
-}).$mount('#app')
+app.use(pinia);
+app.mount("#app");
