@@ -1,116 +1,115 @@
-import { randomRange } from 'Libs/random';
-import { defaultGameState } from 'Libs/gameState';
-import entityFactory from 'Libs/entities/prompts';
-import { findById } from 'Libs/entities';
-import Vue from 'vue';
-
-const state = {
-    ...defaultGameState('actions'),
-};
-
-const getters = {
-    die: (state,) => state.d10 - state.d6,
-    currentRoll: (state, getters) => {
-        if (isNaN(getters.die)) {
-            return '?'
-        } 
-
-        return `${getters.die} (+${state.d10}, -${state.d6})`;
+import { defineStore } from "pinia";
+import { randomRange } from "../lib/random";
+import { defaultGameState } from "../lib/gameState";
+import entityFactory from "../lib/entities/prompts";
+import { findById } from "../lib/entities";
+export const useActionsStore = defineStore("actions", {
+  state: () => ({
+    ...defaultGameState("actions"),
+  }),
+  getters: {
+    die: (state) => state.d10 - state.d6,
+    currentRoll: (state) => {
+      const die = state.d10 - state.d6;
+      if (isNaN(die)) {
+        return "?";
+      }
+      return `${die} (+${state.d10}, -${state.d6})`;
     },
-    prompts: (state) => {
-        const prompts = [...state.prompts];
-
-        prompts.sort((a, b) => a.page > b.page ? -1 : 1);
-
-        return prompts;
+    sortedPrompts: (state) => {
+      const prompts = [...state.prompts];
+      prompts.sort((a, b) => (a.page > b.page ? -1 : 1));
+      return prompts;
     },
     currentPrompt: (state) => state.prompts[state.currentPromptIdx] ?? {},
-};
-
-const mutations = {
-    rollD6: (state) => state.d6 = randomRange(1, 6),
-    setD6: (state, value) => state.d6 = value,
-    setD10: (state, value) => state.d10 = value,
-    rollD10: (state) => state.d10 = randomRange(1, 10),
-    saveRoll: (state, roll) => state.lastRoll = roll,
-    makePrompt: (state, prompt) => state.prompts.push(entityFactory({page: prompt, count: 0})),
-    addPrompt: (state, prompt) => state.prompts.push(entityFactory(prompt)),
-    setPrompts: (state, prompts) => state.prompts = prompts,
-    removePrompt: (state, prompt) => {
-        const found = findById(state.prompts, prompt.id);
-        state.prompts.splice(found.idx, 1);
+  },
+  actions: {
+    rollD6() {
+      this.d6 = randomRange(1, 6);
     },
-    incrementPrompt: (state, prompt) => {
-        const found = findById(state.prompts, prompt.id);
-        Vue.set(found.entity, 'count', found.entity.count + 1)
+    setD6(value) {
+      this.d6 = value;
     },
-    decrementPrompt: (state, prompt) => {
-        const found = findById(state.prompts, prompt.id);
-        Vue.set(found.entity, 'count', found.entity.count - 1)
+    setD10(value) {
+      this.d10 = value;
     },
-    setCurrentPromptIdx: (state, idx) => state.currentPromptIdx = idx,
-}
+    rollD10() {
+      this.d10 = randomRange(1, 10);
+    },
+    saveRoll(roll) {
+      this.lastRoll = roll;
+    },
+    makePrompt(prompt) {
+      this.prompts.push(entityFactory({ page: prompt, count: 0 }));
+    },
+    addPrompt(prompt) {
+      this.prompts.push(entityFactory(prompt));
+    },
+    setPrompts(prompts) {
+      this.prompts = prompts;
+    },
+    removePrompt(prompt) {
+      const found = findById(this.prompts, prompt.id);
 
-const actions = {
-     roll: ({commit, getters, state}) => {
-        commit('saveRoll', getters.currentRoll);
-        commit('rollD6');
-        commit('rollD10');
+      const current = this.currentPrompt;
 
-        let newPrompt = (getters.currentPrompt && getters.currentPrompt.page) ?? 0;
+      this.prompts.splice(found.idx, 1);
 
-        newPrompt += getters.die;
+      const currentFound = findById(this.prompts, current.id);
 
-        if (newPrompt <= 0) {
-            newPrompt = 1;
+      this.currentPromptIdx = currentFound.idx;
+    },
+    incrementPrompt(prompt) {
+      const found = findById(this.prompts, prompt.id);
+      if (found && found.entity) {
+        found.entity.count = (found.entity.count || 0) + 1;
+      }
+    },
+    decrementPrompt(prompt) {
+      const found = findById(this.prompts, prompt.id);
+      if (found && found.entity) {
+        found.entity.count = (found.entity.count || 0) - 1;
+      }
+    },
+    setCurrentPromptIdx(idx) {
+      this.currentPromptIdx = idx;
+    },
+    roll() {
+      this.saveRoll(this.currentRoll);
+      this.rollD6();
+      this.rollD10();
+
+      let newPrompt = (this.currentPrompt && this.currentPrompt.page) ?? 0;
+      newPrompt += this.die;
+      if (newPrompt <= 0) {
+        newPrompt = 1;
+      }
+
+      let promptIdx = this.currentPromptIdx;
+      const promptFound = this.prompts.some((prompt, idx) => {
+        if (prompt.page === newPrompt) {
+          promptIdx = idx;
+          if (prompt.count === 3) {
+            promptIdx += 1;
+            newPrompt += 1;
+          }
+          return true;
         }
-
-        var promptIdx = state.currentPromptIdx;
-
-        const promptFound = state.prompts.some((prompt, idx) => {
-            if (prompt.page === newPrompt) {
-                promptIdx = idx;
-                if (prompt.count === 3) {
-                    promptIdx += 1;
-                    newPrompt += 1;
-                }
-                return true;
-            }
-        });
-
-        if (!promptFound) {
-            promptIdx = state.prompts.length;
-        }
-
-        var prompt = state.prompts[promptIdx];
-
-        if (!prompt) {
-            prompt = entityFactory({page: newPrompt});
-            commit('addPrompt', prompt)
-        }
-
-        commit('incrementPrompt', prompt);
-        commit('setCurrentPromptIdx', promptIdx);
+      });
+      if (!promptFound) {
+        promptIdx = this.prompts.length;
+      }
+      let prompt = this.prompts[promptIdx];
+      if (!prompt) {
+        prompt = entityFactory({ page: newPrompt });
+        this.addPrompt(prompt);
+      }
+      this.incrementPrompt(prompt);
+      this.setCurrentPromptIdx(promptIdx);
     },
-    makePromptCurrent({commit, state}, prompt) {
-        const found = findById(state.prompts, prompt.id);
-        commit('setCurrentPromptIdx', found.idx);
+    makePromptCurrent(prompt) {
+      const found = findById(this.prompts, prompt.id);
+      this.setCurrentPromptIdx(found.idx);
     },
-    removePrompt({commit, getters, state}, prompt) {
-        const currentPrompt = getters.currentPrompt;
-
-        commit('removePrompt', prompt);
-
-        const found = findById(state.prompts, currentPrompt.id);
-
-        commit('setCurrentPromptIdx', found.idx);
-    },
-}
-
-export default {
-    namespaced: true,
-    state,
-    getters,
-    mutations,
-    actions,
-};
+  },
+});
