@@ -1,29 +1,32 @@
-const state = {
-    message: '',
+import { defineStore } from "pinia";
+
+export const useNotificationsStore = defineStore("notifications", {
+  state: () => ({
+    message: "",
     visible: false,
-    type: 'default',
-}
-
-const mutations = {
-    toggle: state => state.visible = !state.visible,
-    hide: state => state.visible = false,
-    show: state => state.visible = true,
-    setMessage: (state, msg) => state.message = msg,
-    setType: (state, type) => state.type = type,
-}
-
-const actions = {
-    showNotification({commit}, {message, type}) {
-        commit('hide');
-        commit('setMessage', message);
-        commit('setType', type)
-        commit('show');
-    }
-};
-
-export default  {
-    namespaced: true,
-    state,
-    mutations,
-    actions,
-}
+    type: "default",
+  }),
+  actions: {
+    toggle() {
+      this.visible = !this.visible;
+    },
+    hide() {
+      this.visible = false;
+    },
+    show() {
+      this.visible = true;
+    },
+    setMessage(msg) {
+      this.message = msg;
+    },
+    setType(type) {
+      this.type = type;
+    },
+    showNotification({ message, type }) {
+      this.hide();
+      this.setMessage(message);
+      this.setType(type);
+      this.show();
+    },
+  },
+});
