@@ -3,13 +3,9 @@ import FormToggleComponent from "Components/FormToggleComponent";
 import { shallowMount, mount } from "@vue/test-utils";
 
 describe("components/FormToggleComponent.vue", () => {
-  it("Has the correct component name", () => {
-    expect(FormToggleComponent.name).toEqual("FormToggleComponent");
-  });
-
   it("Renders a full width button when showControls is false", () => {
-    const wrapper = shallowMount(FormToggleComponent, {
-      propsData: {
+    const wrapper = mount(FormToggleComponent, {
+      props: {
         showControls: false,
       },
       slots: {
@@ -28,7 +24,7 @@ describe("components/FormToggleComponent.vue", () => {
     "Renders a full width button with type %s",
     (type) => {
       const wrapper = shallowMount(FormToggleComponent, {
-        propsData: {
+        props: {
           showControls: false,
           type,
         },
@@ -40,7 +36,7 @@ describe("components/FormToggleComponent.vue", () => {
 
   it("Emits a toggle event when the full width button is clicked", async () => {
     const config = {
-      propsData: {
+      props: {
         showControls: false,
       },
       slots: {
@@ -56,14 +52,12 @@ describe("components/FormToggleComponent.vue", () => {
     await wrapper.vm.$nextTick();
 
     expect(wrapper.emitted().toggle).toBeTruthy();
-    expect(wrapper.emitted().toggle[0][0]).toEqual(
-      config.propsData.showControls
-    );
+    expect(wrapper.emitted().toggle[0][0]).toEqual(config.props.showControls);
   });
 
   it("Renders the form when showControls is true", () => {
-    const wrapper = shallowMount(FormToggleComponent, {
-      propsData: {
+    const wrapper = mount(FormToggleComponent, {
+      props: {
         showControls: true,
       },
       slots: {
@@ -79,7 +73,7 @@ describe("components/FormToggleComponent.vue", () => {
 
   it("Emits a toggle event when the cancel button in the form is clicked", async () => {
     const config = {
-      propsData: {
+      props: {
         showControls: true,
       },
       slots: {
@@ -93,7 +87,7 @@ describe("components/FormToggleComponent.vue", () => {
       .findAllComponents({ name: "ButtonComponent" })
       .filter((w) => w.text() === "Cancel");
 
-    expect(buttons.exists()).toBe(true);
+    expect(buttons.length).toBeGreaterThan(0);
 
     const button = buttons.at(0);
 
@@ -102,14 +96,12 @@ describe("components/FormToggleComponent.vue", () => {
     await wrapper.vm.$nextTick();
 
     expect(wrapper.emitted().toggle).toBeTruthy();
-    expect(wrapper.emitted().toggle[0][0]).toEqual(
-      config.propsData.showControls
-    );
+    expect(wrapper.emitted().toggle[0][0]).toEqual(config.props.showControls);
   });
 
   it("Emits a save event when the save button in the form is clicked", async () => {
     const config = {
-      propsData: {
+      props: {
         showControls: true,
       },
       slots: {
@@ -123,7 +115,7 @@ describe("components/FormToggleComponent.vue", () => {
       .findAllComponents({ name: "ButtonComponent" })
       .filter((w) => w.text() === "Save");
 
-    expect(buttons.exists()).toBe(true);
+    expect(buttons.length).toBeGreaterThan(0);
 
     const button = buttons.at(0);
 
@@ -137,8 +129,8 @@ describe("components/FormToggleComponent.vue", () => {
   it.each([true, false])(
     "Renders controls based on showControls prop",
     (showControls) => {
-      const wrapper = shallowMount(FormToggleComponent, {
-        propsData: {
+      const wrapper = mount(FormToggleComponent, {
+        props: {
           showControls,
         },
         slots: {
@@ -150,7 +142,7 @@ describe("components/FormToggleComponent.vue", () => {
       if (showControls) {
         const form = wrapper.findComponent({ name: "FormComponent" });
         expect(form.exists()).toBe(true);
-        expect(form.text()).toEqual("Form content");
+        expect(form.text()).toContain("Form content");
       } else {
         const button = wrapper.findComponent({ name: "ButtonComponent" });
         expect(button.exists()).toBe(true);
