@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import SaveMenuComponent from "Components/SaveMenuComponent";
 import SlideDownPanelComponent from "Components/SlideDownPanelComponent";
 import ButtonComponent from "Components/ButtonComponent";
-import { shallowMount } from "@vue/test-utils";
+import { shallowMount, mount } from "@vue/test-utils";
 import { getStateFromStore, serialize } from "Libs/gameState";
 import localStorage, { supportsLocalStorage } from "Libs/localStorage";
 
@@ -21,72 +21,52 @@ describe("SaveMenuComponent", () => {
     vi.resetAllMocks();
   });
 
-  it("Has the correct component name", () => {
-    expect(SaveMenuComponent.name).toEqual("SaveMenuComponent");
-  });
-
   it("Renders a SlideDownPanelComponent with a 'Save' heading", () => {
-    const wrapper = shallowMount(SaveMenuComponent, {
-      stubs: {
-        SlideDownPanelComponent,
-      },
-    });
+    const wrapper = mount(SaveMenuComponent);
 
     expect(wrapper.findComponent(SlideDownPanelComponent).exists()).toBe(true);
 
     expect(wrapper.text()).toContain("Save");
   });
 
-  it("Renders ButtonComponents with the correct labels", () => {
-    const wrapper = shallowMount(SaveMenuComponent, {
-      data() {
-        return { saving: true };
-      },
-      stubs: {
-        SlideDownPanelComponent,
-        ButtonComponent,
-      },
-    });
+  it("Renders ButtonComponents with the correct labels", async () => {
+    const wrapper = mount(SaveMenuComponent);
+
+    // Click toggle button to open the panel
+    const toggleButton = wrapper.findAllComponents(ButtonComponent).at(0);
+    await toggleButton.trigger("click");
 
     const buttons = wrapper.findAllComponents(ButtonComponent);
 
-    expect(supportsLocalStorage).toHaveBeenCalled();
+    // SlideDownPanel has 1 button (Show/Close), plus 2 buttons for To File and To Local Storage
     expect(buttons.length).toEqual(3);
-    expect(buttons.at(0).text()).toEqual("Close");
+    // Skip the first button which is from SlideDownPanel
     expect(buttons.at(1).text()).toEqual("To File");
     expect(buttons.at(2).text()).toEqual("To Local Storage");
   });
 
-  it("Does not render the 'To Local Storage' button if local storage is not supported", () => {
+  it("Does not render the 'To Local Storage' button if local storage is not supported", async () => {
     supportsLocalStorage.mockImplementation(() => false);
 
-    const wrapper = shallowMount(SaveMenuComponent, {
-      data() {
-        return { saving: true };
-      },
-      stubs: {
-        SlideDownPanelComponent,
-        ButtonComponent,
-      },
-    });
+    const wrapper = mount(SaveMenuComponent);
+
+    // Click toggle button to open the panel
+    const toggleButton = wrapper.findAllComponents(ButtonComponent).at(0);
+    await toggleButton.trigger("click");
 
     const buttons = wrapper.findAllComponents(ButtonComponent);
 
+    // SlideDownPanel has 1 button (Show/Close), plus 1 button for To File only
     expect(buttons.length).toEqual(2);
-    expect(buttons.at(0).text()).toEqual("Close");
     expect(buttons.at(1).text()).toEqual("To File");
   });
 
   it("Calls 'toFile' when the 'To File' button is clicked", async () => {
-    const wrapper = shallowMount(SaveMenuComponent, {
-      data() {
-        return { saving: true };
-      },
-      stubs: {
-        SlideDownPanelComponent,
-        ButtonComponent,
-      },
-    });
+    const wrapper = mount(SaveMenuComponent);
+
+    // Click toggle button to open the panel
+    const toggleButton = wrapper.findAllComponents(ButtonComponent).at(0);
+    await toggleButton.trigger("click");
 
     const buttons = wrapper
       .findAllComponents(ButtonComponent)
@@ -104,19 +84,14 @@ describe("SaveMenuComponent", () => {
     expect(getStateFromStore).toHaveBeenCalled();
     expect(serialize).toHaveBeenCalledWith({ test: "data" });
     expect(global.URL.createObjectURL).toHaveBeenCalled();
-    expect(wrapper.vm.saving).toBe(false);
   });
 
   it("Calls 'toLocalStorage' when the 'To Local Storage' button is clicked", async () => {
-    const wrapper = shallowMount(SaveMenuComponent, {
-      data() {
-        return { saving: true };
-      },
-      stubs: {
-        SlideDownPanelComponent,
-        ButtonComponent,
-      },
-    });
+    const wrapper = mount(SaveMenuComponent);
+
+    // Click toggle button to open the panel
+    const toggleButton = wrapper.findAllComponents(ButtonComponent).at(0);
+    await toggleButton.trigger("click");
 
     const buttons = wrapper
       .findAllComponents(ButtonComponent)
@@ -132,6 +107,5 @@ describe("SaveMenuComponent", () => {
     expect(getStateFromStore).toHaveBeenCalled();
     expect(serialize).toHaveBeenCalledWith({ test: "data" });
     expect(localStorage.set).toHaveBeenCalledWith("save-game", "serialized");
-    expect(wrapper.vm.saving).toBe(false);
   });
 });
